feat: default layout small code refactor (#225)
* feat: default layout small code refactor

* fix: wrong return example

---------
